Create landing page for legacy docs redirection (#394)
* Add own landing page for legacy docs

* Add HomeLayout component
